azure-deploy/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
python-dotenv==1.0.0
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
qdrant-client==1.16.2
httpx==0.28.1
